backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-docx
PyMuPDF
requests
openai
python-dotenv